requirements.txt
fastapi
uvicorn
gemini_webapi
python-dotenv
black
browser-cookie3
